Move grammar files to parser/ dir
Eliminate the grammar/ dir (this was added before parser/ dir was
introduced in master branch).
