+++ .prettierignore
@@ -4,4 +4,4 @@
 /node_modules
 yarn.lock
 LICENSE
-/src/grammar/grammar.ts
+/src/parser/grammar.ts
