+++ src/parser/LexerAdapter.ts
